pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fedsfd"
version = "0.1.0"
description = "Federated process simulation using Stock-Flow Diagrams discovered from object-centric event logs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pm4py>=2.7.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "jupyter"]

[tool.setuptools.packages.find]
where = ["src"]